1.79 add buses: 100, 14, 33c
git-svn-id: https://minsktranssched.svn.sourceforge.net/svnroot/minsktranssched/trunk@114 1edd68fb-5641-41d5-9a6f-76477202bf19

+++ MinskTransSched/src/resources/ScheduleLoader.java
@@ -110,7 +110,7 @@ public class ScheduleLoader
 	{
 		// load 'dscheds'
 		DataInputStream dis = new DataInputStream(getClass().getResourceAsStream("/dscheds"));
-		int count = dis.readByte();
+		int count = dis.readShort();
 		for (int i = 0; i < count; i++)
 		{
 			// read data
@@ -120,7 +120,7 @@ public class ScheduleLoader
 			int dayTo = dis.readByte();
 			short baseBusStop = dis.readByte();
 			short shift = dis.readByte();
-			
+
 			// src
 			Schedule srcSched = FindSchedule(bus, baseBusStop);
